+++ web-marketplace/src/components/organisms/LoginModal/LoginModal.schema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod';
+
+export const emailFormSchema = z.object({
+  email: z.string().email(),
+});
+
+export type EmailFormSchemaType = z.infer<typeof emailFormSchema>;