+++ web-marketplace/src/components/organisms/RolesForm/RolesForm.schema.ts
@@ -1,14 +1,19 @@
 import { z } from 'zod';
 
-import { addressSchema } from './components/AdminModal/AdminModal.schema';
+import {
+  addressSchema,
+  optionalAddressSchema,
+} from './components/AdminModal/AdminModal.schema';
 import { profileModalSchema } from './components/ProfileModal/ProfileModal.schema';
 
-export const rolesFormSchema = z
-  .object({
+export const rolesFormSchema = (isOnChain: boolean) => {
+  const baseRolesSchema = z.object({
     projectDeveloper: profileModalSchema.nullable().optional(),
     verifier: profileModalSchema.nullable().optional(),
-    admin: addressSchema,
-  })
-  .required({ admin: true });
-
-export type RolesFormSchemaType = z.infer<typeof rolesFormSchema>;
+    admin: isOnChain ? addressSchema : optionalAddressSchema,
+  });
+  return isOnChain
+    ? baseRolesSchema.required({ admin: true })
+    : baseRolesSchema;
+};
+export type RolesFormSchemaType = z.infer<ReturnType<typeof rolesFormSchema>>;