+++ web-marketplace/src/lib/queries/react-query/registry-server/graphql/getProjectByIdQuery/getProjectByIdQuery.constants.ts
@@ -1,5 +1,5 @@
-export const getProjectByIdKey = (handle: string): string[] => [
+export const getProjectByIdKey = (id: string): string[] => [
   'graphql',
   'projectById',
-  handle,
+  id,
 ];