+++ web-storybook/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.14.0](https://github.com/regen-network/regen-web/compare/v1.13.0...v1.14.0) (2023-10-23)
+
+### Features
+
+- add complete tailwind color palette ([#2183](https://github.com/regen-network/regen-web/issues/2183)) ([8eac9b2](https://github.com/regen-network/regen-web/commit/8eac9b2fc52c09bcc1047534c0646ce4ff5256d6))
+- add user account settings page ([#2167](https://github.com/regen-network/regen-web/issues/2167)) ([f1a1d77](https://github.com/regen-network/regen-web/commit/f1a1d77055f3d63f79919badca3a09744feed943))
+
 # [1.13.0](https://github.com/regen-network/regen-web/compare/v1.12.2...v1.13.0) (2023-09-27)
 
 **Note:** Version bump only for package web-storybook